Rename the project so it doesn't match what nuke thinks is a test project
