+++ source/Octopus.Tentacle.Tests/Octopus.Tentacle.Tests.csproj
@@ -28,7 +28,7 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
-    <ProjectReference Include="..\Octopus.Tentacle.Tests.Common\Octopus.Tentacle.Tests.Common.csproj" />
+    <ProjectReference Include="..\Octopus.Tentacle.CommonTestUtils\Octopus.Tentacle.CommonTestUtils.csproj" />
     <ProjectReference Include="..\Octopus.Tentacle\Octopus.Tentacle.csproj" />
     <PackageReference Include="NSubstitute" Version="4.4.0" />
     <PackageReference Include="FluentAssertions" Version="6.7.0" />
